% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{AliasContext}
\alias{AliasContext}
\title{AliasContext Object}
\usage{
AliasContext(name = NULL)
}
\arguments{
\item{name}{The alias name}
}
\value{
AliasContext object
}
\description{
AliasContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An alias to a repo revision.
}

