% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{Breakpoint.labels}
\alias{Breakpoint.labels}
\title{Breakpoint.labels Object}
\usage{
Breakpoint.labels()
}
\value{
Breakpoint.labels object
}
\description{
Breakpoint.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of custom breakpoint properties, populated by the agent, to bedisplayed to the user.
}
\seealso{
Other Breakpoint functions: \code{\link{Breakpoint}}
}
