% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{CloudRepoSourceContext}
\alias{CloudRepoSourceContext}
\title{CloudRepoSourceContext Object}
\usage{
CloudRepoSourceContext(repoId = NULL, revisionId = NULL, aliasName = NULL,
  aliasContext = NULL)
}
\arguments{
\item{repoId}{The ID of the repo}

\item{revisionId}{A revision ID}

\item{aliasName}{The name of an alias (branch, tag, etc}

\item{aliasContext}{An alias, which may be a branch or tag}
}
\value{
CloudRepoSourceContext object
}
\description{
CloudRepoSourceContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A CloudRepoSourceContext denotes a particular revision in a cloud repo (a repo hosted by the Google Cloud Platform).
}

