% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{CloudWorkspaceId}
\alias{CloudWorkspaceId}
\title{CloudWorkspaceId Object}
\usage{
CloudWorkspaceId(repoId = NULL, name = NULL)
}
\arguments{
\item{repoId}{The ID of the repo containing the workspace}

\item{name}{The unique name of the workspace within the repo}
}
\value{
CloudWorkspaceId object
}
\description{
CloudWorkspaceId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A CloudWorkspaceId is a unique identifier for a cloud workspace. A cloud workspace is a place associated with a repo where modified files can be stored before they are committed.
}

