% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_functions.R
\name{controller.debuggees.register}
\alias{controller.debuggees.register}
\title{Registers the debuggee with the controller service.All agents attached to the same application should call this method withthe same request content to get back the same stable `debuggee_id`. Agentsshould call this method again whenever `google.rpc.Code.NOT_FOUND` isreturned from any controller method.This allows the controller service to disable the agent or recover from anydata loss. If the debuggee is disabled by the server, the response willhave `is_disabled` set to `true`.}
\usage{
controller.debuggees.register(RegisterDebuggeeRequest)
}
\arguments{
\item{RegisterDebuggeeRequest}{The \link{RegisterDebuggeeRequest} object to pass to this method
#' @importFrom googleAuthR gar_api_generator}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud_debugger
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud_debugger)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{http://cloud.google.com/debugger}{Google Documentation}

Other RegisterDebuggeeRequest functions: \code{\link{RegisterDebuggeeRequest}}
}
