% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{Debuggee.labels}
\alias{Debuggee.labels}
\title{Debuggee.labels Object}
\usage{
Debuggee.labels()
}
\value{
Debuggee.labels object
}
\description{
Debuggee.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A set of custom debuggee properties, populated by the agent, to bedisplayed to the user.
}
\seealso{
Other Debuggee functions: \code{\link{Debuggee}}
}
