% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_functions.R
\name{debugger.debuggees.list}
\alias{debugger.debuggees.list}
\title{Lists all the debuggees that the user can set breakpoints to.}
\usage{
debugger.debuggees.list(includeInactive = NULL, project = NULL,
  clientVersion = NULL)
}
\arguments{
\item{includeInactive}{When set to `true`, the result includes all debuggees}

\item{project}{Project number of a Google Cloud project whose debuggees to list}

\item{clientVersion}{The client version making the call}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/cloud_debugger
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/cloud_debugger)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{http://cloud.google.com/debugger}{Google Documentation}
}
