% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{ExtendedSourceContext.labels}
\alias{ExtendedSourceContext.labels}
\title{ExtendedSourceContext.labels Object}
\usage{
ExtendedSourceContext.labels()
}
\value{
ExtendedSourceContext.labels object
}
\description{
ExtendedSourceContext.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Labels with user defined metadata.
}
\seealso{
Other ExtendedSourceContext functions: \code{\link{ExtendedSourceContext}}
}

