% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{ExtendedSourceContext}
\alias{ExtendedSourceContext}
\title{ExtendedSourceContext Object}
\usage{
ExtendedSourceContext(ExtendedSourceContext.labels = NULL, labels = NULL,
  context = NULL)
}
\arguments{
\item{ExtendedSourceContext.labels}{The \link{ExtendedSourceContext.labels} object or list of objects}

\item{labels}{Labels with user defined metadata}

\item{context}{Any source context}
}
\value{
ExtendedSourceContext object
}
\description{
ExtendedSourceContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An ExtendedSourceContext is a SourceContext combined with additionaldetails describing the context.
}
\seealso{
Other ExtendedSourceContext functions: \code{\link{ExtendedSourceContext.labels}}
}
