% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{FormatMessage}
\alias{FormatMessage}
\title{FormatMessage Object}
\usage{
FormatMessage(parameters = NULL, format = NULL)
}
\arguments{
\item{parameters}{Optional parameters to be embedded into the message}

\item{format}{Format template for the message}
}
\value{
FormatMessage object
}
\description{
FormatMessage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a message with parameters.
}
