% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{GerritSourceContext}
\alias{GerritSourceContext}
\title{GerritSourceContext Object}
\usage{
GerritSourceContext(hostUri = NULL, revisionId = NULL, aliasName = NULL,
  gerritProject = NULL, aliasContext = NULL)
}
\arguments{
\item{hostUri}{The URI of a running Gerrit instance}

\item{revisionId}{A revision (commit) ID}

\item{aliasName}{The name of an alias (branch, tag, etc}

\item{gerritProject}{The full project name within the host}

\item{aliasContext}{An alias, which may be a branch or tag}
}
\value{
GerritSourceContext object
}
\description{
GerritSourceContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A SourceContext referring to a Gerrit project.
}
