% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{GetBreakpointResponse}
\alias{GetBreakpointResponse}
\title{GetBreakpointResponse Object}
\usage{
GetBreakpointResponse(breakpoint = NULL)
}
\arguments{
\item{breakpoint}{Complete breakpoint state}
}
\value{
GetBreakpointResponse object
}
\description{
GetBreakpointResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for getting breakpoint information.
}
