% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{GitSourceContext}
\alias{GitSourceContext}
\title{GitSourceContext Object}
\usage{
GitSourceContext(url = NULL, revisionId = NULL)
}
\arguments{
\item{url}{Git repository URL}

\item{revisionId}{Git commit hash}
}
\value{
GitSourceContext object
}
\description{
GitSourceContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A GitSourceContext denotes a particular revision in a third party Git repository (e.g. GitHub).
}

