% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{ListActiveBreakpointsResponse}
\alias{ListActiveBreakpointsResponse}
\title{ListActiveBreakpointsResponse Object}
\usage{
ListActiveBreakpointsResponse(breakpoints = NULL, nextWaitToken = NULL,
  waitExpired = NULL)
}
\arguments{
\item{breakpoints}{List of all active breakpoints}

\item{nextWaitToken}{A wait token that can be used in the next method call to block until the list of breakpoints changes}

\item{waitExpired}{The `wait_expired` field is set to true by the server when the request times out and the field `success_on_timeout` is set to true}
}
\value{
ListActiveBreakpointsResponse object
}
\description{
ListActiveBreakpointsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for listing active breakpoints.
}

