% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{ListBreakpointsResponse}
\alias{ListBreakpointsResponse}
\title{ListBreakpointsResponse Object}
\usage{
ListBreakpointsResponse(breakpoints = NULL, nextWaitToken = NULL)
}
\arguments{
\item{breakpoints}{List of breakpoints matching the request}

\item{nextWaitToken}{A wait token that can be used in the next call to `list` (REST) or}
}
\value{
ListBreakpointsResponse object
}
\description{
ListBreakpointsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for listing breakpoints.
}
