% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{ListDebuggeesResponse}
\alias{ListDebuggeesResponse}
\title{ListDebuggeesResponse Object}
\usage{
ListDebuggeesResponse(debuggees = NULL)
}
\arguments{
\item{debuggees}{List of debuggees accessible to the calling user}
}
\value{
ListDebuggeesResponse object
}
\description{
ListDebuggeesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for listing debuggees.
}

