% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{ProjectRepoId}
\alias{ProjectRepoId}
\title{ProjectRepoId Object}
\usage{
ProjectRepoId(projectId = NULL, repoName = NULL)
}
\arguments{
\item{projectId}{The ID of the project}

\item{repoName}{The name of the repo}
}
\value{
ProjectRepoId object
}
\description{
ProjectRepoId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Selects a repo using a Google Cloud Platform project ID (e.g. winged-cargo-31) and a repo name within that project.
}

