% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{RegisterDebuggeeRequest}
\alias{RegisterDebuggeeRequest}
\title{RegisterDebuggeeRequest Object}
\usage{
RegisterDebuggeeRequest(debuggee = NULL)
}
\arguments{
\item{debuggee}{Debuggee information to register}
}
\value{
RegisterDebuggeeRequest object
}
\description{
RegisterDebuggeeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to register a debuggee.
}
\seealso{
Other RegisterDebuggeeRequest functions: \code{\link{controller.debuggees.register}}
}
