% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{RegisterDebuggeeResponse}
\alias{RegisterDebuggeeResponse}
\title{RegisterDebuggeeResponse Object}
\usage{
RegisterDebuggeeResponse(debuggee = NULL)
}
\arguments{
\item{debuggee}{Debuggee resource}
}
\value{
RegisterDebuggeeResponse object
}
\description{
RegisterDebuggeeResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for registering a debuggee.
}

