% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{RepoId}
\alias{RepoId}
\title{RepoId Object}
\usage{
RepoId(projectRepoId = NULL, uid = NULL)
}
\arguments{
\item{projectRepoId}{A combination of a project ID and a repo name}

\item{uid}{A server-assigned, globally unique identifier}
}
\value{
RepoId object
}
\description{
RepoId Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A unique identifier for a cloud repo.
}
