% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{SetBreakpointResponse}
\alias{SetBreakpointResponse}
\title{SetBreakpointResponse Object}
\usage{
SetBreakpointResponse(breakpoint = NULL)
}
\arguments{
\item{breakpoint}{Breakpoint resource}
}
\value{
SetBreakpointResponse object
}
\description{
SetBreakpointResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for setting a breakpoint.
}
