% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{StatusMessage}
\alias{StatusMessage}
\title{StatusMessage Object}
\usage{
StatusMessage(isError = NULL, refersTo = NULL, description = NULL)
}
\arguments{
\item{isError}{Distinguishes errors from informational messages}

\item{refersTo}{Reference to which the message applies}

\item{description}{Status message text}
}
\value{
StatusMessage object
}
\description{
StatusMessage Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a contextual status message. The message can indicate an error or informational status, and refer to specific parts of the containing object. For example, the `Breakpoint.status` field can indicate an error referring to the `BREAKPOINT_SOURCE_LOCATION` with the message `Location not found`.
}

