% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{UpdateActiveBreakpointRequest}
\alias{UpdateActiveBreakpointRequest}
\title{UpdateActiveBreakpointRequest Object}
\usage{
UpdateActiveBreakpointRequest(breakpoint = NULL)
}
\arguments{
\item{breakpoint}{Updated breakpoint information}
}
\value{
UpdateActiveBreakpointRequest object
}
\description{
UpdateActiveBreakpointRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to update an active breakpoint.
}

