% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{UpdateActiveBreakpointResponse}
\alias{UpdateActiveBreakpointResponse}
\title{UpdateActiveBreakpointResponse Object}
\usage{
UpdateActiveBreakpointResponse()
}
\value{
UpdateActiveBreakpointResponse object
}
\description{
UpdateActiveBreakpointResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response for updating an active breakpoint.The message is defined to allow future extensions.
}
