% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{CloudWorkspaceSourceContext}
\alias{CloudWorkspaceSourceContext}
\title{CloudWorkspaceSourceContext Object}
\usage{
CloudWorkspaceSourceContext(snapshotId = NULL, workspaceId = NULL)
}
\arguments{
\item{snapshotId}{The ID of the snapshot}

\item{workspaceId}{The ID of the workspace}
}
\value{
CloudWorkspaceSourceContext object
}
\description{
CloudWorkspaceSourceContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A CloudWorkspaceSourceContext denotes a workspace at a particular snapshot.
}
