% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{Debuggee}
\alias{Debuggee}
\title{Debuggee Object}
\usage{
Debuggee(Debuggee.labels = NULL, project = NULL, id = NULL,
  agentVersion = NULL, isDisabled = NULL, description = NULL,
  uniquifier = NULL, sourceContexts = NULL, extSourceContexts = NULL,
  labels = NULL, status = NULL, isInactive = NULL)
}
\arguments{
\item{Debuggee.labels}{The \link{Debuggee.labels} object or list of objects}

\item{project}{Project the debuggee is associated with}

\item{id}{Unique identifier for the debuggee generated by the controller service}

\item{agentVersion}{Version ID of the agent release}

\item{isDisabled}{If set to `true`, indicates that the agent should disable itself and}

\item{description}{Human readable description of the debuggee}

\item{uniquifier}{Debuggee uniquifier within the project}

\item{sourceContexts}{References to the locations and revisions of the source code used in the}

\item{extSourceContexts}{References to the locations and revisions of the source code used in the}

\item{labels}{A set of custom debuggee properties, populated by the agent, to be}

\item{status}{Human readable message to be displayed to the user about this debuggee}

\item{isInactive}{If set to `true`, indicates that the debuggee is considered as inactive by}
}
\value{
Debuggee object
}
\description{
Debuggee Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents the application to debug. The application may include one or morereplicated processes executing the same code. Each of these processes isattached with a debugger agent, carrying out the debugging commands.The agents attached to the same debuggee are identified by using exactly thesame field values when registering.
}
\seealso{
Other Debuggee functions: \code{\link{Debuggee.labels}}
}
