% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{RegisterDebuggeeRequest}
\alias{RegisterDebuggeeRequest}
\title{Stackdriver Debugger API Objects 
Examines the call stack and variables of a running application without stopping or slowing it down.}
\usage{
RegisterDebuggeeRequest(debuggee = NULL)
}
\arguments{
\item{debuggee}{Debuggee information to register}
}
\value{
RegisterDebuggeeRequest object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 23:17:00
filename: /Users/mark/dev/R/autoGoogleAPI/googleclouddebuggerv2.auto/R/clouddebugger_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
RegisterDebuggeeRequest Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request to register a debuggee.
}
\seealso{
Other RegisterDebuggeeRequest functions: \code{\link{controller.debuggees.register}}
}

