% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{SourceContext}
\alias{SourceContext}
\title{SourceContext Object}
\usage{
SourceContext(cloudRepo = NULL, cloudWorkspace = NULL, gerrit = NULL,
  git = NULL)
}
\arguments{
\item{cloudRepo}{A SourceContext referring to a revision in a cloud repo}

\item{cloudWorkspace}{A SourceContext referring to a snapshot in a cloud workspace}

\item{gerrit}{A SourceContext referring to a Gerrit project}

\item{git}{A SourceContext referring to any third party Git repo (e}
}
\value{
SourceContext object
}
\description{
SourceContext Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A SourceContext is a reference to a tree of files. A SourceContext together with a path point to a unique revision of a single file or directory.
}

