% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{SourceLocation}
\alias{SourceLocation}
\title{SourceLocation Object}
\usage{
SourceLocation(path = NULL, line = NULL)
}
\arguments{
\item{path}{Path to the source file within the source context of the target binary}

\item{line}{Line inside the file}
}
\value{
SourceLocation object
}
\description{
SourceLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a location in the source code.
}

