% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{StackFrame}
\alias{StackFrame}
\title{StackFrame Object}
\usage{
StackFrame(function. = NULL, location = NULL, arguments = NULL,
  locals = NULL)
}
\arguments{
\item{function.}{Demangled function name at the call site}

\item{location}{Source location of the call site}

\item{arguments}{Set of arguments passed to this function}

\item{locals}{Set of local variables at the stack frame location}
}
\value{
StackFrame object
}
\description{
StackFrame Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Represents a stack frame context.
}

