% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clouddebugger_objects.R
\name{Variable}
\alias{Variable}
\title{Variable Object}
\usage{
Variable(name = NULL, value = NULL, type = NULL, members = NULL,
  varTableIndex = NULL, status = NULL)
}
\arguments{
\item{name}{Name of the variable, if any}

\item{value}{Simple value of the variable}

\item{type}{Variable type (e}

\item{members}{Members contained or pointed to by the variable}

\item{varTableIndex}{Reference to a variable in the shared variable table}

\item{status}{Status associated with the variable}
}
\value{
Variable object
}
\description{
Variable Object
}

